Add files via upload
This project was a test back in my "Algorithm and Programming Logic" class at the university right in the first semester. We use a tool called "Portugol", that is a pseudolanguage that allow us write codes in portuguese so we could learn the fundamentals easily. It contains a document "QUIZ.txt" with all of the necessary informations to run the game, such as: "Question | Topic | Right Answer | Ans1 | Ans2 | Ans3 | Ans4", in this format. Must be greater than 5 questions to the game runs free. The document "txt" must be in the same directory as the Portugol Quiz "N2 Quiz.por".